ENH: Remove domain from variables
Now when you need to run the playbook there are no edits that need to be made as all information is stored in the ansible vault. Also, by removing the domain we can run the playbook against multiple hosts at once (dev and prod) grafana. The domain is replaced with the inventory host. This means when using the hosts file we must supply domain names that would match the certificate