DOC: Ignore package-latest lint errors
We want to use the latest package versions to ensure we are always up-to-date with security fixes. When the packages version upgrade breaks the service we will fix it. The playbook will almost always be run in development before production so any problems from new package versions should be detected early on.