ENH: Create a terraform playbook
This role and playbook will deploy the terraform project. This means we can avoid using a bash script which is complicated to make idempotent and users only have to deploy with one tool, in this case ansible. Rather than deploying with 2 tools. Also we can use Jinja to create the hosts file rather than bash which is much clearer.