BUG: Fix ansible user module bad behaviour
When providing the user module with "group" it removes all existing
groups of the user and adds only that one group. We change this to
"groups" and use "append" to add the group to the other groups the user
is already a part of. This makes it more similar to "usermod -aG group
user"