ENH: Moving to a dynamic inventory
Using the openstack plugin to create a dynamic inventory at runtime based off of metadata on the VM. To achieve this we add a service and env tag to the VM metadata allowing them to be put into groups. We can also set up multiple deployments now. Meaning we can run production and development at the same time.