MAINT: Changing how IPs are found
Now we have a dynamic inventory we access the IPs of hosts in a different way. Adding this lovely complicated maps everywhere to produce the same result. In the dynamic inventory the "inventory_hostname" variable is the hostname of the VM. Whereas when we were templating our own hosts.ini file we put in the IP addresses of the VMs. The "inventory_hostname" cannot be changed to IP address as the plugin allows it to be only the actual VM hostname or UUID