BUG: Fix Alertmanager and Prometheus Changes
When handlers were introduced to Alertmanager and Prometheus they were
not indented correctly. Fixing this by de-indenting all notify
statements by 2 spaces.

Also fixes the Copy module not being able to create the parent to its
destination directory. We solve this by creating the service
(/opt/service) directory before moving the binaries. We then change the
unarchive module to check for the executable binary instead of the
folder itself. So it checks for /opt/service/service rather than for
/opt/service.