BUG: Certbot was not renewing certificates
There was no cron job to stop HAProxy before trying to renew. Therefore certificates were not being renewed and close to expiry. We could fix this by adding another cron job to stop the service or edit the cerbot renew job to do it aswell. Instead we are merging all the cron jobs into one script and using a single cron job to run the script. The script will update certbot, stop haproxy, renew the cert, copy the cert and restart haproxy