MAINT: Remove volume facts from play
We cannot rely on the OpenStack API to provide accurate information about where the device is mounted. This is due to a discrepancy between OpenStack and the OS of the VM. To work around this we will try the two most common devices, sdb and vdb. If both of these fail intervention will be required. A bonus to this is it removes the convoluted ansible fact solution previously used.